+++ client/manga/src/components/MangaCard.jsx
@@ -8,8 +8,11 @@ import { faTrashAlt } from '@fortawesome/free-solid-svg-icons';
 const MangaCard = ({ manga ,onDelete}) => {
     const {user} = useContext(AuthContext)
     const deleteManga = (event) => {
-        event.preventDefault(); // Prevent default form submission
-        onDelete(manga._id); // Call the onDelete function with manga id
+        event.preventDefault(); 
+        const confirmDelete = window.confirm("Are you sure to delete this manga?");
+        if(confirmDelete){
+            onDelete(manga._id); 
+        }
     };
     return (
         <Link to={`/manga/${manga._id}`} key={manga._id} className="block group">
